ai/server/requirements.txt
uvicorn
fastapi
nltk
numpy
torchtext
pandas
pinecone-client
python-dotenv
